requirements.txt
beautifulsoup4
ftfy
google-api-python-client
google-cloud-language
wand
#libmagickwand
facebook-sdk
